Move blog e2e tests into separate file
